+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Module `robodyno.robots` was missing the `ThreeDofCartesian` class.
+
 ## [1.6.1] - 2023-05-29
 
 ### Added